requirements.txt
fastapi[standard]>=0.115.5
pydantic>=2.4.2
sqlalchemy>=2.0.36
uvicorn[standard]>=0.32.1
python-dotenv